Documents/000AI相关/工作编程项目/自动化脚本-抓web小红书/requirements.txt
playwright>=1.40.0
pandas>=2.2.0
pyyaml>=6.0.1
pytesseract>=0.3.10
Pillow>=10.1.0